# Go_Server/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个集合了多种语言的Web靶场，该系统是以食谱菜单管理系统为场景去编写，一种实战化形式的安全漏洞靶场，其中存在多个安全漏洞，需要我们去探索和发现。\\n\\n该项目旨在帮助安全研究人员和爱好者了解和掌握关于不同语言系统的渗透测试和代码审计知识。",
        "title": "食谱菜单管理系统靶场",
        "contact": {
            "name": "Vulnerabilities_Server",
            "url": "https://github.com/A7cc/Vulnerabilities_Server"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/settings/backupsdb": {
            "get": {
                "tags": [
                    "鉴权接口-系统设置相关方法"
                ],
                "summary": "备份数据库",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/settings/deletedb": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "鉴权接口-系统设置相关方法"
                ],
                "summary": "删除备份数据库",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "删除数据库备份文件",
                        "name": "dbfile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/settings/downdb": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "鉴权接口-系统设置相关方法"
                ],
                "summary": "数据库下载",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "下载数据库备份文件",
                        "name": "dbfile",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/settings/getdb": {
            "post": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "鉴权接口-系统设置相关方法"
                ],
                "summary": "获取备份数据库列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "数据库备份目录",
                        "name": "dir",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/user/add": {
            "post": {
                "tags": [
                    "鉴权接口-用户相关方法"
                ],
                "summary": "新增用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "新增用户信息",
                        "name": "AddUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.AddUserRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/delete/{id}": {
            "delete": {
                "tags": [
                    "鉴权接口-用户相关方法"
                ],
                "summary": "删除用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "删除用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/detail": {
            "get": {
                "tags": [
                    "鉴权接口-用户相关方法"
                ],
                "summary": "获取管理员详细信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "获取用户ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/get": {
            "get": {
                "tags": [
                    "鉴权接口-用户相关方法"
                ],
                "summary": "获取用户列表信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "获取用户查询参数",
                        "name": "GetUserListRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.GetUserListRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/update": {
            "put": {
                "tags": [
                    "鉴权接口-用户相关方法"
                ],
                "summary": "修改用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "更新用户信息参数",
                        "name": "UpdateUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/define.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
...
        "define.AddUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer"
                },
                "sex": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "define.GetUserListRequest": {
            "type": "object",
            "properties": {
                "keyword": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "define.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "role_id": {
                    "type": "integer"
                },
                "sex": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}